refactor: drop the dead self-signing path and stale certs docs
GenerateServerCert has been superseded by (*CA).IssueServerCert; it and its
tests are gone, the two test helpers that still called it now issue from a
CA, and what remains of selfsigned.go is shared PKI material, so the file is
renamed to pki.go. The package comment no longer describes Phase 1
self-signing as current, and state.AuditEntry.Detail is documented as the
short fixed reason it actually carries.

Item 6 of #46.
